refactor : EventForm
