+++ packages/assignment-7/src/App.tsx
@@ -1,17 +1,9 @@
-import { ChangeEvent, useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import {
   Alert,
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogOverlay,
   AlertIcon,
   AlertTitle,
   Box,
-  Button,
-  Checkbox,
   CloseButton,
   Flex,
   FormControl,
@@ -27,10 +19,8 @@ import {
   Text,
   Th,
   Thead,
-  Tooltip,
   Tr,
   useInterval,
-  useToast,
   VStack,
 } from "@chakra-ui/react";
 import {
@@ -46,11 +36,11 @@ import {
   getDaysInMonth,
   getWeekDates,
 } from "./utils/date-utils";
-import { useClosure } from "./components/alert/hooks/useClosure";
-import { useSaveEvent } from "./components/event/hooks/useSaveEvent";
 import { useGetEvent } from "./components/event/hooks/useGetEvent";
+import { useDeleteEvent } from "./components/event/hooks/useDeleteEvent";
+import EventForm from "./components/event/EventForm";
 
-type RepeatType = "none" | "daily" | "weekly" | "monthly" | "yearly";
+export type RepeatType = "none" | "daily" | "weekly" | "monthly" | "yearly";
 
 interface RepeatInfo {
   type: RepeatType;
@@ -70,8 +60,6 @@ export interface Event {
   repeat: RepeatInfo;
   notificationTime: number; // 분 단위로 저장
 }
-
-const categories = ["업무", "개인", "가족", "기타"];
 
 const weekDays = ["일", "월", "화", "수", "목", "금", "토"];
 
@@ -110,171 +98,25 @@ const fetchHolidays = (year: number, month: number) => {
 function App() {
   const { events, fetchEvents } = useGetEvent();
 
-  const [title, setTitle] = useState("");
-  const [date, setDate] = useState("");
-  const [startTime, setStartTime] = useState("");
-  const [endTime, setEndTime] = useState("");
-  const [description, setDescription] = useState("");
-  const [location, setLocation] = useState("");
-  const [category, setCategory] = useState("");
   const [view, setView] = useState<"week" | "month">("month");
-  const [isRepeating, setIsRepeating] = useState(false);
-  const [repeatType, setRepeatType] = useState<RepeatType>("none");
-  const [repeatInterval, setRepeatInterval] = useState(1);
-  const [repeatEndDate, setRepeatEndDate] = useState("");
-  const [notificationTime, setNotificationTime] = useState(10);
   const [notifications, setNotifications] = useState<
     { id: number; message: string }[]
   >([]);
   const [notifiedEvents, setNotifiedEvents] = useState<number[]>([]);
-
-  const {
-    isOpen: isOverlapDialogOpen,
-    onOpen: onOpenOverlapDialog,
-    onClose: onCloseOverlapDialog,
-  } = useClosure();
-
-  const [overlappingEvents, setOverlappingEvents] = useState<Event[]>([]);
-
-  const [startTimeError, setStartTimeError] = useState<string | null>(null);
-  const [endTimeError, setEndTimeError] = useState<string | null>(null);
 
   const [currentDate, setCurrentDate] = useState(new Date());
 
   const [searchTerm, setSearchTerm] = useState("");
   const [holidays, setHolidays] = useState<{ [key: string]: string }>({});
 
-  const cancelRef = useRef<HTMLButtonElement>(null);
+  // 수정 중 이벤트
+  const [editingEvent, setEditingEvent] = useState<Event | null>(null);
 
-  const toast = useToast();
-
-  const { saveEvent, editingEvent, setEditingEvent } = useSaveEvent();
-
-  const addOrUpdateEvent = async () => {
-    if (!title || !date || !startTime || !endTime) {
-      toast({
-        title: "필수 정보를 모두 입력해주세요.",
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-      return;
-    }
-
-    validateTime(startTime, endTime);
-    if (startTimeError || endTimeError) {
-      toast({
-        title: "시간 설정을 확인해주세요.",
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-      return;
-    }
-
-    const eventData: Event = {
-      id: editingEvent ? editingEvent.id : Date.now(),
-      title,
-      date,
-      startTime,
-      endTime,
-      description,
-      location,
-      category,
-      repeat: {
-        type: isRepeating ? repeatType : "none",
-        interval: repeatInterval,
-        endDate: repeatEndDate || undefined,
-      },
-      notificationTime,
-    };
-
-    const overlapping = findOverlappingEvents(eventData);
-
-    if (overlapping.length > 0) {
-      setOverlappingEvents(overlapping);
-      onOpenOverlapDialog();
-    } else {
-      await saveEvent(eventData);
-      await fetchEvents();
-      resetForm();
-    }
+  const editEvent = (event: Event) => {
+    setEditingEvent(event);
   };
 
-  // const saveEvent = async (eventData: Event) => {
-  //   try {
-  //     let response;
-  //     if (editingEvent) {
-  //       response = await fetch(`/api/events/${eventData.id}`, {
-  //         method: "PUT",
-  //         headers: {
-  //           "Content-Type": "application/json",
-  //         },
-  //         body: JSON.stringify(eventData),
-  //       });
-  //     } else {
-  //       response = await fetch("/api/events", {
-  //         method: "POST",
-  //         headers: {
-  //           "Content-Type": "application/json",
-  //         },
-  //         body: JSON.stringify(eventData),
-  //       });
-  //     }
-
-  //     if (!response.ok) {
-  //       throw new Error("Failed to save event");
-  //     }
-
-  //     await fetchEvents(); // 이벤트 목록 새로고침
-  //     setEditingEvent(null);
-  //     resetForm();
-  //     toast({
-  //       title: editingEvent
-  //         ? "일정이 수정되었습니다."
-  //         : "일정이 추가되었습니다.",
-  //       status: "success",
-  //       duration: 3000,
-  //       isClosable: true,
-  //     });
-  //   } catch (error) {
-  //     console.error("Error saving event:", error);
-  //     toast({
-  //       title: "일정 저장 실패",
-  //       status: "error",
-  //       duration: 3000,
-  //       isClosable: true,
-  //     });
-  //   }
-  // };
-
-  const deleteEvent = async (id: number) => {
-    try {
-      const response = await fetch(`/api/events/${id}`, {
-        method: "DELETE",
-      });
-
-      if (!response.ok) {
-        throw new Error("Failed to delete event");
-      }
-
-      await fetchEvents(); // 이벤트 목록 새로고침
-      toast({
-        title: "일정이 삭제되었습니다.",
-        status: "info",
-        duration: 3000,
-        isClosable: true,
-      });
-    } catch (error) {
-      console.error("Error deleting event:", error);
-      toast({
-        title: "일정 삭제 실패",
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-    }
-  };
+  const { deleteEvent } = useDeleteEvent(fetchEvents);
 
   const checkUpcomingEvents = async () => {
     const now = new Date();
@@ -302,86 +144,6 @@ function App() {
         console.error("Error updating notification status:", error);
       }
     }
-  };
-
-  const validateTime = (start: string, end: string) => {
-    if (!start || !end) return;
-
-    const startDate = new Date(`2000-01-01T${start}`);
-    const endDate = new Date(`2000-01-01T${end}`);
-
-    if (startDate >= endDate) {
-      setStartTimeError("시작 시간은 종료 시간보다 빨라야 합니다.");
-      setEndTimeError("종료 시간은 시작 시간보다 늦어야 합니다.");
-    } else {
-      setStartTimeError(null);
-      setEndTimeError(null);
-    }
-  };
-
-  const handleStartTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newStartTime = e.target.value;
-    setStartTime(newStartTime);
-    validateTime(newStartTime, endTime);
-  };
-
-  const handleEndTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newEndTime = e.target.value;
-    setEndTime(newEndTime);
-    validateTime(startTime, newEndTime);
-  };
-
-  // 날짜 문자열을 Date 객체로 변환하는 함수
-  const parseDateTime = (date: string, time: string): Date => {
-    return new Date(`${date}T${time}`);
-  };
-
-  // 두 일정이 겹치는지 확인하는 함수
-  const isOverlapping = (event1: Event, event2: Event): boolean => {
-    const start1 = parseDateTime(event1.date, event1.startTime);
-    const end1 = parseDateTime(event1.date, event1.endTime);
-    const start2 = parseDateTime(event2.date, event2.startTime);
-    const end2 = parseDateTime(event2.date, event2.endTime);
-
-    return start1 < end2 && start2 < end1;
-  };
-
-  // 겹치는 일정을 찾는 함수
-  const findOverlappingEvents = (newEvent: Event): Event[] => {
-    return events.filter(
-      (event) => event.id !== newEvent.id && isOverlapping(event, newEvent)
-    );
-  };
-
-  const resetForm = () => {
-    setTitle("");
-    setDate("");
-    setStartTime("");
-    setEndTime("");
-    setDescription("");
-    setLocation("");
-    setCategory("");
-    setEditingEvent(null);
-    setIsRepeating(false);
-    setRepeatType("none");
-    setRepeatInterval(1);
-    setRepeatEndDate("");
-  };
-
-  const editEvent = (event: Event) => {
-    setEditingEvent(event);
-    setTitle(event.title);
-    setDate(event.date);
-    setStartTime(event.startTime);
-    setEndTime(event.endTime);
-    setDescription(event.description);
-    setLocation(event.location);
-    setCategory(event.category);
-    setIsRepeating(event.repeat.type !== "none");
-    setRepeatType(event.repeat.type);
-    setRepeatInterval(event.repeat.interval);
-    setRepeatEndDate(event.repeat.endDate || "");
-    setNotificationTime(event.notificationTime);
   };
 
   const navigate = (direction: "prev" | "next") => {
@@ -604,165 +366,14 @@ function App() {
     setHolidays(newHolidays);
   }, [currentDate]);
 
-  // useEffect(() => {
-  //   fetchEvents();
-  // }, []);
-
   return (
     <Box w="full" h="100vh" m="auto" p={5}>
       <Flex gap={6} h="full">
-        {/* form 영역 */}
-        <VStack w="400px" spacing={5} align="stretch">
-          <Heading>{editingEvent ? "일정 수정" : "일정 추가"}</Heading>
-
-          <FormControl>
-            <FormLabel>제목</FormLabel>
-            <Input value={title} onChange={(e) => setTitle(e.target.value)} />
-          </FormControl>
-
-          <FormControl>
-            <FormLabel>날짜</FormLabel>
-            <Input
-              type="date"
-              value={date}
-              onChange={(e) => setDate(e.target.value)}
-            />
-          </FormControl>
-
-          <HStack width="100%">
-            <FormControl>
-              <FormLabel>시작 시간</FormLabel>
-              <Tooltip
-                label={startTimeError}
-                isOpen={!!startTimeError}
-                placement="top"
-              >
-                <Input
-                  type="time"
-                  value={startTime}
-                  onChange={handleStartTimeChange}
-                  onBlur={() => validateTime(startTime, endTime)}
-                  isInvalid={!!startTimeError}
-                />
-              </Tooltip>
-            </FormControl>
-            <FormControl>
-              <FormLabel>종료 시간</FormLabel>
-              <Tooltip
-                label={endTimeError}
-                isOpen={!!endTimeError}
-                placement="top"
-              >
-                <Input
-                  type="time"
-                  value={endTime}
-                  onChange={handleEndTimeChange}
-                  onBlur={() => validateTime(startTime, endTime)}
-                  isInvalid={!!endTimeError}
-                />
-              </Tooltip>
-            </FormControl>
-          </HStack>
-
-          <FormControl>
-            <FormLabel>설명</FormLabel>
-            <Input
-              value={description}
-              onChange={(e) => setDescription(e.target.value)}
-            />
-          </FormControl>
-
-          <FormControl>
-            <FormLabel>위치</FormLabel>
-            <Input
-              value={location}
-              onChange={(e) => setLocation(e.target.value)}
-            />
-          </FormControl>
-
-          <FormControl>
-            <FormLabel>카테고리</FormLabel>
-            <Select
-              value={category}
-              onChange={(e) => setCategory(e.target.value)}
-            >
-              <option value="">카테고리 선택</option>
-              {categories.map((cat) => (
-                <option key={cat} value={cat}>
-                  {cat}
-                </option>
-              ))}
-            </Select>
-          </FormControl>
-
-          <FormControl>
-            <FormLabel>반복 설정</FormLabel>
-            <Checkbox
-              isChecked={isRepeating}
-              onChange={(e) => setIsRepeating(e.target.checked)}
-            >
-              반복 일정
-            </Checkbox>
-          </FormControl>
-
-          <FormControl>
-            <FormLabel>알림 설정</FormLabel>
-            <Select
-              value={notificationTime}
-              onChange={(e) => setNotificationTime(Number(e.target.value))}
-            >
-              {notificationOptions.map((option) => (
-                <option key={option.value} value={option.value}>
-                  {option.label}
-                </option>
-              ))}
-            </Select>
-          </FormControl>
-
-          {isRepeating && (
-            <VStack width="100%">
-              <FormControl>
-                <FormLabel>반복 유형</FormLabel>
-                <Select
-                  value={repeatType}
-                  onChange={(e) => setRepeatType(e.target.value as RepeatType)}
-                >
-                  <option value="daily">매일</option>
-                  <option value="weekly">매주</option>
-                  <option value="monthly">매월</option>
-                  <option value="yearly">매년</option>
-                </Select>
-              </FormControl>
-              <HStack width="100%">
-                <FormControl>
-                  <FormLabel>반복 간격</FormLabel>
-                  <Input
-                    type="number"
-                    value={repeatInterval}
-                    onChange={(e) => setRepeatInterval(Number(e.target.value))}
-                    min={1}
-                  />
-                </FormControl>
-                <FormControl>
-                  <FormLabel>반복 종료일</FormLabel>
-                  <Input
-                    type="date"
-                    value={repeatEndDate}
-                    onChange={(e) => setRepeatEndDate(e.target.value)}
-                  />
-                </FormControl>
-              </HStack>
-            </VStack>
-          )}
-
-          <Button
-            data-testid="event-submit-button"
-            onClick={addOrUpdateEvent}
-            colorScheme="blue"
-          >
-            {editingEvent ? "일정 수정" : "일정 추가"}
-          </Button>
-        </VStack>
+        <EventForm
+          events={events}
+          fetchEvents={fetchEvents}
+          editingEvent={editingEvent}
+        />
 
         <VStack flex={1} spacing={5} align="stretch">
           <Heading>일정 보기</Heading>
@@ -878,63 +489,6 @@ function App() {
           )}
         </VStack>
       </Flex>
-
-      {/* 일정 겹침 알람 dialog */}
-      <AlertDialog
-        isOpen={isOverlapDialogOpen}
-        leastDestructiveRef={cancelRef}
-        onClose={() => onCloseOverlapDialog()}
-      >
-        <AlertDialogOverlay>
-          <AlertDialogContent>
-            <AlertDialogHeader fontSize="lg" fontWeight="bold">
-              일정 겹침 경고
-            </AlertDialogHeader>
-
-            <AlertDialogBody>
-              다음 일정과 겹칩니다:
-              {overlappingEvents.map((event) => (
-                <Text key={event.id}>
-                  {event.title} ({event.date} {event.startTime}-{event.endTime})
-                </Text>
-              ))}
-              계속 진행하시겠습니까?
-            </AlertDialogBody>
-
-            <AlertDialogFooter>
-              <Button ref={cancelRef} onClick={() => onCloseOverlapDialog()}>
-                취소
-              </Button>
-              <Button
-                colorScheme="red"
-                onClick={async () => {
-                  onCloseOverlapDialog();
-                  await saveEvent({
-                    id: editingEvent ? editingEvent.id : Date.now(),
-                    title,
-                    date,
-                    startTime,
-                    endTime,
-                    description,
-                    location,
-                    category,
-                    repeat: {
-                      type: isRepeating ? repeatType : "none",
-                      interval: repeatInterval,
-                      endDate: repeatEndDate || undefined,
-                    },
-                    notificationTime,
-                  });
-                  await fetchEvents();
-                }}
-                ml={3}
-              >
-                계속 진행
-              </Button>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialogOverlay>
-      </AlertDialog>
 
       {notifications.length > 0 && (
         <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
