+++ packages/assignment-7/src/main.tsx
@@ -4,9 +4,9 @@ import App from "./App.tsx";
 import { ChakraProvider } from "@chakra-ui/react";
 
 async function prepare() {
-  const { setupWorker } = await import("msw/browser");
-  const { handlers } = await import("../mock/handler.ts");
-  const worker = setupWorker(...handlers);
+  // const { setupWorker } = await import("msw/browser");
+  // const { handlers } = await import("../mock/handler.ts");
+  // const worker = setupWorker(...handlers);
   // return worker.start();
 }
