+++ packages/assignment-7/src/utils/date-utils.ts
@@ -1,3 +1,5 @@
+import { Event } from "../App";
+
 /**
  * 특정 년도와 달을 통해 일수를 얻는 함수
  * @param year 현재 년도
@@ -59,8 +61,43 @@ const getCurrentDate = (): string => {
 
   const dateString = year + "-" + month + "-" + day;
 
-  console.log(dateString);
   return dateString;
+};
+
+/**
+ * 날짜 문자열을 Date 객체로 변환하는 함수
+ * @param date
+ * @param time
+ * @returns
+ */
+const parseDateTime = (date: string, time: string): Date => {
+  return new Date(`${date}T${time}`);
+};
+
+/**
+ * 두 일정이 겹치는지 확인하는 함수
+ * @param event1
+ * @param event2
+ * @returns
+ */
+const isOverlapping = (event1: Event, event2: Event): boolean => {
+  const start1 = parseDateTime(event1.date, event1.startTime);
+  const end1 = parseDateTime(event1.date, event1.endTime);
+  const start2 = parseDateTime(event2.date, event2.startTime);
+  const end2 = parseDateTime(event2.date, event2.endTime);
+
+  return start1 < end2 && start2 < end1;
+};
+
+/**
+ * 겹치는 일정을 찾는 함수
+ * @param newEvent
+ * @returns
+ */
+const findOverlappingEvents = (oldEvent: Event[], newEvent: Event): Event[] => {
+  return oldEvent.filter(
+    (event) => event.id !== newEvent.id && isOverlapping(event, newEvent)
+  );
 };
 
 export {
@@ -69,4 +106,7 @@ export {
   formatWeek,
   formatMonth,
   getCurrentDate,
+  parseDateTime,
+  isOverlapping,
+  findOverlappingEvents,
 };
